remove dropdown if not needed
