+++ webapp/lib/f_tracking/widgets/tracking_page.dart
@@ -8,12 +8,11 @@ import 'package:cosmos_notifier/f_home/widgets/subwidgets/bottom_navigation_bar_
 import 'package:cosmos_notifier/f_home/widgets/subwidgets/footer_widget.dart';
 import 'package:cosmos_notifier/f_tracking/services/state/tracker_row.dart';
 import 'package:cosmos_notifier/f_tracking/services/tracker_provider.dart';
+import 'package:cosmos_notifier/f_tracking/widgets/subwidgets/address_input_widget.dart';
 import 'package:cosmos_notifier/style.dart';
-import 'package:debounce_throttle/debounce_throttle.dart';
 import 'package:fixnum/fixnum.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:riverpod_messages/riverpod_messages.dart';
 
@@ -152,10 +151,9 @@ class TrackingPage extends StatelessWidget {
                         icon: const Icon(Icons.arrow_downward),
                         iconSize: iconSizeSmall,
                         elevation: 16,
-                        style: const TextStyle(color: Colors.deepPurple),
+                        style: const TextStyle(),
                         underline: Container(
                           height: 2,
-                          color: Colors.deepPurpleAccent,
                         ),
                         onChanged: (TrackerChatRoom? newValue) async {
                           if (newValue == null) {
@@ -231,7 +229,7 @@ class TrackingPage extends StatelessWidget {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  const HeaderWidget(),
+                  HeaderWidget(false),
                   const SizedBox(height: 10),
                   table(context),
                   validationError(context),
@@ -244,32 +242,6 @@ class TrackingPage extends StatelessWidget {
         ],
       ),
       bottomNavigationBar: BottomNavigationBarWidget(jwtManager: jwtManager),
-    );
-  }
-}
-
-class AddressInputWidget extends HookWidget {
-  final WidgetRef ref;
-  final TrackerRow trackerRow;
-
-  const AddressInputWidget(this.ref, this.trackerRow, {Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final debouncer = Debouncer<String>(const Duration(milliseconds: 200), initialValue: "");
-    debouncer.values.listen((value) {
-      ref.read(trackerNotifierProvider.notifier).updateTracker(trackerRow.copyWith(address: value));
-    });
-    final controller = useTextEditingController(text: trackerRow.shortenedAddress);
-    return TextField(
-      controller: controller,
-      decoration: const InputDecoration(
-        hintText: 'Enter wallet address',
-        border: InputBorder.none,
-        enabledBorder: InputBorder.none,
-        hintStyle: TextStyle(color: Colors.grey, fontSize: 12),
-      ),
-      onChanged: (value) => {debouncer.value = value},
     );
   }
 }
